AllInOne8.java splitted on AllInOne7.java and AllInOne8.java.
